move query methods in on file
